+++ lib/src/generator/broker/broker_api_endpoint.dart
@@ -1,24 +1,39 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:cl_datahub/src/generator/broker/endpoint_param.dart';
 
 import '../transfer_bean/field_type.dart';
 
 class BrokerApiEndpoint {
   final String name;
-  final FieldType? payloadType;
+  final List<EndpointParam> params;
   final FieldType? replyType;
-  final String? payloadName;
-  final bool isAsync;
 
   bool get isRpc => replyType != null;
 
+  String get returnTypeStatement =>
+      isRpc ? 'Future<${replyType!.typeName}>' : 'void';
+
   BrokerApiEndpoint(
     this.name,
-    this.payloadName,
-    this.payloadType,
+    this.params,
     this.replyType,
-    this.isAsync,
   );
+
+  factory BrokerApiEndpoint.fromMethod(MethodElement m) {
+    final params = m.parameters.map((p) => EndpointParam.fromParam(p)).toList();
+    final replyType = findReplyType(m);
+    if (!endpointIsAsyncOrVoid(m)) {
+      throw Exception('Endpoint return types must be of type Future. '
+          '${m.name} has a non-Future return type.');
+    }
+
+    return BrokerApiEndpoint(
+      m.name,
+      params,
+      replyType,
+    );
+  }
 }
 
 FieldType? findPayloadType(MethodElement m) {
@@ -60,6 +75,8 @@ FieldType? findReplyType(MethodElement m) {
   return FieldType.fromDartType(m.returnType);
 }
 
-bool endpointIsAsync(MethodElement m) {
-  return m.returnType.isDartAsyncFuture || m.returnType.isDartAsyncFutureOr;
+bool endpointIsAsyncOrVoid(MethodElement m) {
+  return m.returnType.isDartAsyncFuture ||
+      m.returnType.isDartAsyncFutureOr ||
+      m.returnType.isVoid;
 }