+++ lib/src/generator/broker/broker_api_service_generator.dart
@@ -23,20 +23,10 @@ class BrokerApiServiceGenerator extends GeneratorForAnnotation<BrokerApi> {
     final queueName = annotation.read('queueName').literalValue as String;
     final durable = annotation.read('durable').literalValue as bool;
 
-    final endpoints = classElement.methods.where((m) => !m.isPrivate).map((m) {
-      final payloadName = findPayloadName(m);
-      final payloadType = findPayloadType(m);
-      final replyType = findReplyType(m);
-      final isAsync = endpointIsAsync(m);
-
-      return BrokerApiEndpoint(
-        m.name,
-        payloadName,
-        payloadType,
-        replyType,
-        isAsync,
-      );
-    }).toList();
+    final endpoints = classElement.methods
+        .where((m) => !m.isPrivate)
+        .map((m) => BrokerApiEndpoint.fromMethod(m))
+        .toList();
 
     var result = BrokerApiServiceBuilder(
       classElement.name,